small token fixes

+++ src/commands/ban.ts
@@ -7,7 +7,7 @@ import { getGuildData } from "../guildData";
 export class BanCommand extends Command {
     info: CommandInfo = {
         name: "ban",
-        syntax: [["user", "member"], ["string", "reason", false], ["float", "delaySec", false]],
+        syntax: [["user", "member"], ["string", "reason", false], ["int", "delaySec", false]],
         description: "банит участника сервера `member`",
         permission: "admin",
         group: "Администрирование"
